Offload far-distance calculation to CanSee class
