+++ src/components/Utilities/CanSee.ts
@@ -1,4 +1,5 @@
 import { BresenhamIterator } from './BresenhamIterator';
+import { GameState } from '../Builder/Types/GameState';
 import { Map } from '../MapModel/Types/Map';
 import { Mob } from '../Mobs/Mob';
 import { WorldPoint } from '../MapModel/WorldPoint';
@@ -115,6 +116,51 @@ export class CanSee {
     b: WorldPoint,
     threshold: number,
   ): boolean {
-    return a.squaredDistanceTo(b) < threshold;
+    return a.squaredDistanceTo(b) <= threshold;
+  }
+
+  /**
+   * Calculates the far distance for visibility based on player position, map, and game state.
+   *
+   * @param {WorldPoint} playerPos - The position of the player on the map.
+   * @param {Map} map - The current map.
+   * @param {GameState} game - The game state containing player stats.
+   * @return {number} The calculated far distance for visibility.
+   */
+  public static getFarDist(
+    playerPos: WorldPoint,
+    map: Map,
+    game: GameState,
+  ): number {
+    const glowRange = 10;
+    const maxVisibilityRange = 75;
+    let farDist = game.stats.currentVisRange || 50;
+
+    const glowingRocks = this.countLightSources(playerPos, map, glowRange);
+    farDist *= Math.pow(2, glowingRocks);
+
+    return Math.min(farDist, maxVisibilityRange);
+  }
+
+  /**
+   * Counts the number of light sources in the vicinity of the player position within a specified diameter.
+   *
+   * @param {WorldPoint} playerPos - The position of the player on the map.
+   * @param {Map} map - The current map.
+   * @param {number} diameter - The diameter within which to count light sources.
+   * @return {number} The count of light sources in the specified vicinity.
+   */
+  private static countLightSources(
+    playerPos: WorldPoint,
+    map: Map,
+    diameter: number,
+  ): number {
+    let lightSources = 0;
+    for (const neighbor of playerPos.getNeighbors(diameter * 0.5)) {
+      if (map.isLegalPoint(neighbor) && map.cell(neighbor).isGlowing()) {
+        lightSources++;
+      }
+    }
+    return lightSources;
   }
 }
